Move functions to utils file, rename and update imports
