+++ packages/nextjs/app/lending/components/modals/RepayTransactionModal.tsx.tsx
@@ -10,7 +10,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Address } from "viem";
 import { useChainId } from "wagmi";
 import { useTranslation } from "~~/app/context/LanguageContext";
-import { getBlockExplorerUrl } from "~~/app/utils/blockExplorer";
+import { getBlockExplorerUrl } from "~~/app/utils/utils";
 import useRepay from "~~/hooks/useRepay";
 
 interface ModalProps {
