+++ packages/nextjs/app/lending/components/modals/WithdrawTransactionModal.tsx
@@ -8,7 +8,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Address } from "viem";
 import { useChainId } from "wagmi";
 import { useTranslation } from "~~/app/context/LanguageContext";
-import { getBlockExplorerUrl } from "~~/app/utils/blockExplorer";
+import { getBlockExplorerUrl } from "~~/app/utils/utils";
 import useWithdraw from "~~/hooks/useWithdraw";
 
 interface ModalProps {
